requirements.txt
fastapi==0.60.1
jaeger-client>=4.1.0
opentracing==2.3.0
opentracing-instrumentation==3.1.1
psycopg2-binary==2.8.4
pydantic==1.5.1
python-dotenv==0.12.0
requests==2.23.0
sqlalchemy==1.3.16
starlette==0.13.6
tornado==5.1.1
uvicorn==0.11.5
python-jose==3.1.0
pandas==1.0.3
cmake==3.17.1
numpy==1.18.5
geog==0.0.2
celery==4.4.5
pylint==2.5.2
coverage==5.1

# Geo
GeoAlchemy2==0.7.0
geog==0.0.2
shapely==1.7.0
h3==3.4.1
geojson==2.5.0

# Redis
redis==3.4.1


# test requirements
pytest
pytest-env
pytest-cov
